requirements.txt
atlassian-python-api
requests_ntlm
requests
flask